fix(typescript): move type definitions to match sources
