+++ src/helpers/overrides.d.ts
@@ -1,0 +1,23 @@
+import * as React from 'react';
+import { StyleObject } from 'styletron-react';
+import { Theme } from '../styles';
+
+type StyleOverride<T> =
+  | StyleObject
+  | ((props: { $theme: Theme } & React.PropsWithChildren<T>) => StyleObject);
+
+type ComponentOverride<T> = React.ComponentType<T> | React.RefForwardingComponent<T>;
+
+interface OverrideObject<T> {
+  component?: ComponentOverride<T>;
+  props?: any;
+  style?: StyleOverride<T>;
+}
+
+export type Override<T> = OverrideObject<T> | React.ComponentType<T>;
+
+export interface Overrides<T> {
+  [key: string]: Override<T>;
+}
+
+export function mergeOverrides<T>(target?: Overrides<T>, source?: Overrides<T>): Overrides<T>;
